+++ samples/add-and-drop-new-database-role.js
@@ -1,4 +1,4 @@
-// Copyright 2022 Google LLC
+// Copyright 2024 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -30,19 +30,18 @@ function main(
   // const instanceId = 'my-instance';
   // const databaseId = 'my-database';
   // const projectId = 'my-project-id';
-  // Imports the Google Cloud Spanner client library
+
+  // Imports the Google Cloud client library
   const {Spanner} = require('@google-cloud/spanner');
 
-  // Instantiates a client
+  // creates a client
   const spanner = new Spanner({
     projectId: projectId,
   });
 
-  async function addAndDropNewDatabaseRole() {
-    // Gets a reference to a Cloud Spanner instance and database.
-    const instance = spanner.instance(instanceId);
-    const database = instance.database(databaseId);
+  const databaseAdminClient = spanner.getDatabaseAdminClient();
 
+  async function addAndDropNewDatabaseRole() {
     // Creates a new user defined role and grant permissions
     try {
       const request = [
@@ -51,7 +50,14 @@ function main(
         'CREATE ROLE child',
         'GRANT ROLE parent TO ROLE child',
       ];
-      const [operation] = await database.updateSchema(request);
+      const [operation] = await databaseAdminClient.updateDatabaseDdl({
+        database: databaseAdminClient.databasePath(
+          projectId,
+          instanceId,
+          databaseId
+        ),
+        statements: request,
+      });
 
       console.log('Waiting for operation to complete...');
       await operation.promise();
@@ -65,7 +71,14 @@ function main(
     // A role can't be dropped until all its permissions are revoked.
     try {
       const request = ['REVOKE ROLE parent FROM ROLE child', 'DROP ROLE child'];
-      const [operation] = await database.updateSchema(request);
+      const [operation] = await databaseAdminClient.updateDatabaseDdl({
+        database: databaseAdminClient.databasePath(
+          projectId,
+          instanceId,
+          databaseId
+        ),
+        statements: request,
+      });
 
       console.log('Waiting for operation to complete...');
       await operation.promise();
@@ -74,8 +87,9 @@ function main(
     } catch (err) {
       console.error('ERROR:', err);
     } finally {
-      // Close the database when finished.
-      await database.close();
+      // Close the spanner client when finished.
+      // The databaseAdminClient does not require explicit closure. The closure of the Spanner client will automatically close the databaseAdminClient.
+      spanner.close();
     }
   }
   addAndDropNewDatabaseRole();