+++ samples/pg-sequence-alter.js
@@ -1,4 +1,4 @@
-// Copyright 2023 Google LLC
+// Copyright 2024 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -36,14 +36,20 @@ async function main(instanceId, databaseId, projectId) {
   });
 
   async function alterSequence(instanceId, databaseId) {
-    // Gets a reference to a Cloud Spanner instance and database
-    const instance = spanner.instance(instanceId);
-    const database = instance.database(databaseId);
+    // Gets a reference to a Cloud Spanner Database Admin Client object
+    const databaseAdminClient = spanner.getDatabaseAdminClient();
 
     const request = ['ALTER SEQUENCE Seq SKIP RANGE 1000 5000000'];
 
     try {
-      const [operation] = await database.updateSchema(request);
+      const [operation] = await databaseAdminClient.updateDatabaseDdl({
+        database: databaseAdminClient.databasePath(
+          projectId,
+          instanceId,
+          databaseId
+        ),
+        statements: request,
+      });
 
       console.log('Waiting for operation to complete...');
       await operation.promise();
@@ -54,6 +60,11 @@ async function main(instanceId, databaseId, projectId) {
     } catch (err) {
       console.error('ERROR:', err);
     }
+
+    // Gets a reference to a Cloud Spanner instance and database
+    const instance = spanner.instance(instanceId);
+    const database = instance.database(databaseId);
+
     database.runTransaction(async (err, transaction) => {
       if (err) {
         console.error(err);
@@ -79,8 +90,9 @@ async function main(instanceId, databaseId, projectId) {
       } catch (err) {
         console.error('ERROR:', err);
       } finally {
-        // Close the database when finished.
-        await database.close();
+        // Close the spanner client when finished.
+        // The databaseAdminClient does not require explicit closure. The closure of the Spanner client will automatically close the databaseAdminClient.
+        spanner.close();
       }
     });
   }