+++ src/index.ts
@@ -46,7 +46,13 @@ import {
   CreateInstanceConfigCallback,
   CreateInstanceConfigResponse,
 } from './instance-config';
-import {grpc, GrpcClientOptions, CallOptions, GoogleError} from 'google-gax';
+import {
+  grpc,
+  GrpcClientOptions,
+  CallOptions,
+  GoogleError,
+  ClientOptions,
+} from 'google-gax';
 import {google, google as instanceAdmin} from '../protos/protos';
 import {
   PagedOptions,
@@ -316,7 +322,8 @@ class Spanner extends GrpcService {
       baseUrl:
         options.apiEndpoint ||
         options.servicePath ||
-        v1.SpannerClient.servicePath,
+        // TODO: for TPC, this needs to support universeDomain
+        'spanner.googleapis.com',
       protosDir: path.resolve(__dirname, '../protos'),
       protoServices: {
         Operations: {
@@ -344,6 +351,54 @@ class Spanner extends GrpcService {
       [CLOUD_RESOURCE_HEADER]: this.projectFormattedName_,
     };
     this.directedReadOptions = directedReadOptions;
+  }
+
+  /**
+   * Gets the InstanceAdminClient object.
+   * The returned InstanceAdminClient object is a shared, managed instance and should not be manually closed.
+   * @returns {v1.InstanceAdminClient} The InstanceAdminClient object
+   * @example
+   *  ```
+   * const {Spanner} = require('@google-cloud/spanner');
+   * const spanner = new Spanner({
+   *    projectId: projectId,
+   *  });
+   * const instanceAdminClient = spanner.getInstanceAdminClient();
+   * ```
+   */
+  getInstanceAdminClient(): v1.InstanceAdminClient {
+    const clientName = 'InstanceAdminClient';
+    if (!this.clients_.has(clientName)) {
+      this.clients_.set(
+        clientName,
+        new v1[clientName](this.options as ClientOptions)
+      );
+    }
+    return this.clients_.get(clientName)! as v1.InstanceAdminClient;
+  }
+
+  /**
+   * Gets the DatabaseAdminClient object.
+   * The returned DatabaseAdminClient object is a managed, shared instance and should not be manually closed.
+   * @returns {v1.DatabaseAdminClient} The DatabaseAdminClient object.
+   * @example
+   * ```
+   * const {Spanner} = require('@google-cloud/spanner');
+   * const spanner = new Spanner({
+   *    projectId: projectId,
+   * });
+   * const databaseAdminClient = spanner.getDatabaseAdminClient();
+   * ```
+   */
+  getDatabaseAdminClient(): v1.DatabaseAdminClient {
+    const clientName = 'DatabaseAdminClient';
+    if (!this.clients_.has(clientName)) {
+      this.clients_.set(
+        clientName,
+        new v1[clientName](this.options as ClientOptions)
+      );
+    }
+    return this.clients_.get(clientName)! as v1.DatabaseAdminClient;
   }
 
   /** Closes this Spanner client and cleans up all resources used by it. */
@@ -1740,6 +1795,8 @@ promisifyAll(Spanner, {
     'pgJsonb',
     'operation',
     'timestamp',
+    'getInstanceAdminClient',
+    'getDatabaseAdminClient',
   ],
 });