+++ src/transaction.ts
@@ -69,6 +69,7 @@ export interface RequestOptions {
 export interface CommitOptions {
   requestOptions?: Pick<IRequestOptions, 'priority'>;
   returnCommitStats?: boolean;
+  maxCommitDelay?: spannerClient.protobuf.IDuration;
   gaxOptions?: CallOptions;
 }
 
@@ -979,7 +980,7 @@ export class Snapshot extends EventEmitter {
    *     execution statistics for the SQL statement that
    *     produced this result set.
    * @property {string} partitionToken The partition token.
-   * @property {number} seqno The Sequence number.
+   * @property {number} seqno The Sequence number. This option is used internally and will be overridden.
    * @property {string} sql The SQL string.
    * @property {google.spanner.v1.ExecuteSqlRequest.IQueryOptions} [queryOptions]
    *     Default query options to use with the database. These options will be
@@ -1299,7 +1300,10 @@ export class Snapshot extends EventEmitter {
     if (!is.empty(typeMap)) {
       Object.keys(typeMap).forEach(param => {
         const type = typeMap[param];
-        paramTypes[param] = codec.createTypeObject(type);
+        const typeObject = codec.createTypeObject(type);
+        if (typeObject.code !== 'TYPE_CODE_UNSPECIFIED') {
+          paramTypes[param] = codec.createTypeObject(type);
+        }
       });
     }
 
@@ -1798,6 +1802,9 @@ export class Transaction extends Dml {
    *     with the commit request.
    * @property {boolean} returnCommitStats Include statistics related to the
    *     transaction in the {@link CommitResponse}.
+   * @property {spannerClient.proto.IDuration} maxCommitDelay Maximum amount
+   *     of delay the commit is willing to incur in order to improve
+   *     throughput. Value should be between 0ms and 500ms.
    * @property {object} [gaxOptions] The request configuration options,
    *     See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions}
    *     for more details.
@@ -1887,6 +1894,12 @@ export class Transaction extends Dml {
       (options as CommitOptions).returnCommitStats
     ) {
       reqOpts.returnCommitStats = (options as CommitOptions).returnCommitStats;
+    }
+    if (
+      'maxCommitDelay' in options &&
+      (options as CommitOptions).maxCommitDelay
+    ) {
+      reqOpts.maxCommitDelay = (options as CommitOptions).maxCommitDelay;
     }
     reqOpts.requestOptions = Object.assign(
       requestOptions || {},